Initial Workings
Added in the CSV files for each of the three dataframes. We may need to edit these because there is something funky going on in the average weighted monthly returns one. I also started writing some of the preliminary code for the data cleaning, but it might be nice to go over stuff together.